+++ {{cookiecutter.project_slug}}/backend/app/app/crud/subitem.py
