+++ {{cookiecutter.project_slug}}/backend/app/app/models/item.py
@@ -4,9 +4,11 @@ from sqlalchemy.orm import relationship
 from app.db.base_class import Base
 
 
-class SubItem(Base):
+class Item(Base):
     id = Column(Integer, primary_key=True, index=True)
     title = Column(String, index=True)
     description = Column(String, index=True)
-    item_id = Column(Integer, ForeignKey("item.id"))
-    item = relationship("Item", back_populates="subitems")
+
+    owner_id = Column(Integer, ForeignKey("user.id"))
+    owner = relationship("User", back_populates="items")
+    subitems = relationship("SubItem", back_populates="item")
