+++ {{cookiecutter.project_slug}}/backend/app/app/models/subitem.py
@@ -4,10 +4,10 @@ from sqlalchemy.orm import relationship
 from app.db.base_class import Base
 
 
-class Item(Base):
+class SubItem(Base):
     id = Column(Integer, primary_key=True, index=True)
     title = Column(String, index=True)
     description = Column(String, index=True)
-    owner_id = Column(Integer, ForeignKey("user.id"))
-    owner = relationship("User", back_populates="items")
-    subitems = relationship("SubItem", back_populates="item")
+
+    item_id = Column(Integer, ForeignKey("item.id"))
+    item = relationship("Item", back_populates="subitems")
