+++ {{cookiecutter.project_slug}}/backend/app/app/schemas/item.py
@@ -1,10 +1,12 @@
 from pydantic import BaseModel
 
+from .user import User
 
 # Shared properties
 class ItemBase(BaseModel):
     title: str = None
     description: str = None
+    owner_id: int = None
 
     class Config:
         orm_mode = True
@@ -13,6 +15,7 @@ class ItemBase(BaseModel):
 # Properties to receive on item creation
 class ItemCreate(ItemBase):
     title: str
+    owner_id: int
 
 
 # Properties to receive on item update
@@ -24,14 +27,10 @@ class ItemUpdate(ItemBase):
 class ItemInDBBase(ItemBase):
     id: int
     title: str
-    owner_id: int
+
+    owner: User
 
 
 # Properties to return to client
 class Item(ItemInDBBase):
-    pass
-
-
-# Properties properties stored in DB
-class ItemInDB(ItemInDBBase):
     pass
