+++ {{cookiecutter.project_slug}}/backend/app/app/schemas/subitem.py
@@ -5,7 +5,8 @@ from .item import Item
 
 # Shared properties
 class SubItemBase(BaseModel):
-    name: str = None
+    title: str = None
+    description: str = None
     item_id: int
 
     class Config:
@@ -14,7 +15,7 @@ class SubItemBase(BaseModel):
 
 # Properties to receive on item creation
 class SubItemCreate(SubItemBase):
-    name: str
+    title: str
 
 
 # Properties to receive on item update
@@ -25,14 +26,8 @@ class SubItemUpdate(SubItemBase):
 # Properties shared by models stored in DB
 class SubItemInDBBase(SubItemBase):
     id: int
-    name: str
 
 
 # Properties to return to client
 class SubItem(SubItemInDBBase):
     item : Item
-
-
-# Properties properties stored in DB
-class SubItemInDB(SubItemInDBBase):
-    pass
